Example

Example Julia package repo.

ClassicAlgorithmsCollections.binary_pivot_search — Method

binary_pivot_search(array::Array{Int64,1}, target::Int64)

The idea is to find the pivot point for finding the target in an unsorted array. For this reason, the array has to be divided into two subarrays; a binary search is performed on the subarrays.

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

target::Int64: Target-value to find the position

...

Examples

julia> arr = [2, 5, 4, 7, 2, 8, 9, 3, 10, 2]
julia> target = 3
julia> ClassicAlgorithmsCollections.binary_pivot_search(arr, target)
8

ClassicAlgorithmsCollections.binary_search — Method

binary_search(array::Array{Int64,1}, target::Int64)

The binary search algorithm (BSA) finds a target value's position within a sorted array by using a half-interval cut per each cycle. Thus, the BSA compares the target value to the value of the array's middle element. In the case of inequality, the half array-piece in which the target cannot be will be erased. Next, the search continues on the remaining half array-piece and starts taking the middle element to compare it to the target value. This procedure has to be continued until the target value is found. The search may have to be stopped with a remaining empty half array-piece; consequently, the target is not in the array. For more information see: https://en.wikipedia.org/wiki/Binarysearchalgorithm

...

Arguments

array::Array{Int64,1}: Sorted array of integers

target::Int64: Target-value to find the position

...

Examples

julia> arr = [10,11, 12, 14, 16, 18, 19, 20, 21, 22, 23, 24, 33, 35, 42, 47]
julia> target = 12
julia> ClassicAlgorithmsCollections.binary_search(arr, target)
3

ClassicAlgorithmsCollections.boogle_word_check — Method

boogle_word_check(
    graph::Dict{Int64,Array{String,1}},
    reference_words::Array{String,1},
)

For finding words (reference_words) in a field of chars, the boogle word check algorithm goes for every single char up and down to see if the sum of the chars build a word contained in th refernce word list. For this porpose the Depth First Traversal algorithm in function find_word is used.

Arguments

graph::Dict{Int64,Array{String,1}}: Graph of the connected nodes of chars, which can build the words

reference_words::Array{String,1}: Reference words to search for

Examples

julia> import ClassicAlgorithmsCollections
julia> word_list = ["GEEKS", "FOR", "QUIZ", "GO"]
julia> graph_boogle = Dict(
        1 => ["G", "I", "Z"],
        2 => ["U", "E", "K"],
        3 => ["Q", "S", "E"],
        4 => ["D", "O", "P"],
        5 => ["F", "O", "R"],
        )
julia> ClassicAlgorithmsCollections.boogle_word_check(graph_boogle, word_list) ==
["GEEKS", "QUIZ", "FOR"]]

ClassicAlgorithmsCollections.breadth_first_search — Method

breadth_first_search(graph::Dict{Int64,Array{Int64,1}}, start::Int64)

The breadth-first search (BFS) is an algorithm dedicated to traversing or searching for tree or graph data structures. It starts at a specified tree root (start) for exploring all connected neighbor nodes. The important feature is that the BFS automatically leaves the present depth and passes on to the next nodes at a deeper level. BFS is queue-based. For more information see: https://en.wikipedia.org/wiki/Breadth-first_search

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

start::Int64: Startpoint (first selected vertex) of the graph-traveling process

...

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = Dict(1=> [2, 3], 2=> [3], 3=> [1, 4], 4=> [4])
julia> ClassicAlgorithmsCollections.breadth_first_search(graph, 3)
[3, 1, 4, 2]

ClassicAlgorithmsCollections.bridge_check — Method

bridge_check( graph::Dict{Int64,Array{Int64,1}}, u::Int64, result::Array{Int64,1}, visited::Array{Bool,1}, parent::Array{Int64,1}, low::Array{Int64,1}, disc::Array{Int64,1}, time::Int64, )

For finding a bridge belonging to the current vertex, the algorithm has to remove firstly one by one all edges. Next, the algorithm has to see if the removal of an edge causes a disconnected graph. If yes, for the current pair of u and v, are a bridge between the endpoints is found.

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes

u::Int64: Next to visit vertex

result::Array{Int64,1}: Results of the graph bridges; start and endpoint of the bridge

visited::Array{Bool,1}: Visited vertex

parent::Array{Int64,1}: Parent vertices in DFS tree'

low::Array{Int64,1}: Lowest vertex reachable from subtree

disc::Array{Int64,1}: Discovery time of the visited vertex

time::Int64: Current time

ClassicAlgorithmsCollections.bubble_sorting — Method

bubble_sorting(array::Array{Int64,1})

The Bubble sorting algorithm (BSA) is a primitive sorting algorithm that repeatedly steps through the array by using a double for-loop with n and n-1 size. During the walkthrough, the BSA compares adjacent elements and swaps wrong ordered elements until the array is sorted. For more information see: https://en.wikipedia.org/wiki/Bubble_sort

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

...

Examples

julia> arr = [64, 34, 25, 12, 22, 11, 90] 
julia> ClassicAlgorithmsCollections.bubble_sorting(arr)
[11, 12, 22, 25, 34, 64, 90]

ClassicAlgorithmsCollections.closest_pair_searching — Method

closest_pair_searching(array::Array{Int64,1}, target::Int64)

Finding the closet pair of values for a given target in a sorted array. For this reason, the distance between a pair of of values has to be minimazied with respect to the target. For more information see: https://en.wikipedia.org/wiki/Closestpairofpointsproblem

...

Arguments

array::Array{Int64,1}: Sorted array of integers

target::Int64: Target-value to find the position

...

Examples

julia> arr = [10, 22, 28, 29, 30, 40]
julia> target = 56
julia> ClassicAlgorithmsCollections.closest_pair_searching(arr, target)
(28, 29)

ClassicAlgorithmsCollections.depth_first_search — Method

depth_first_search(graph::Dict{Int64,Array{Int64,1}}, start::Int64)

The depth-first search (DFS) is an algorithm dedicated to traversing or searching for tree or graph data structures. It starts at a specified tree root (start) for exploring as far as possible at each branch. Afterthat the BFS starts automatically backtracking. DFS is stack-based. For more information see: https://en.wikipedia.org/wiki/Depth-first_search

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

start::Int64: Startpoint (first selected vertex) of the graph-traveling process

...

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = Dict(1=> [2, 3], 2=> [3], 3=> [1, 4], 4=> [4])
julia> ClassicAlgorithmsCollections.breadth_first_search(graph, 3)
[3, 1, 2, 4]

ClassicAlgorithmsCollections.dfs_recrusive — Method

dfs_recrusive(
    graph::Dict{Int64,Array{Int64,1}},
    start::Int64,
    visited::Array{Bool,1},
    solution::Array{Int64,1},
)

The recursively call of the dfs_recrusive is essential for exploring each single branch of the graph.

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

start::Int64: Startpoint (first selected vertex) of the graph-traveling process

visited::Array{Bool,1}: Visited nodes of the graph

solution::Array{Int64,1}: Solution of the raph-traveling proces

...

ClassicAlgorithmsCollections.find_global_maximum — Method

find_global_maximum(graph::Dict{Int64,Array{Int64,1}})

Find the total global maximum based on a comparsion between the intial vertex (global_maximum = 0), the current dictionary key, and the accesible vertexes from the array-list (value).

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

...

ClassicAlgorithmsCollections.find_global_maximum_complex — Method

find_global_maximum_complex(graph::Dict{Int64,Array{Tuple{Int64,Int64},1}})

find_global_maximum_complex is similar to find_global_maximum instead is desigend for handling graphs with weights.

...

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes with weights

...

ClassicAlgorithmsCollections.find_parent_in_cycle — Method

find_parent_in_cycle(parent::Array{Int64,1}, i::Int64, ref_key::Int64)

The find parent algorithm is a recursive function to find the subeset of an item i for graph test of being cycling.

Arguments

parent::Array{Int64,1}: Array of the subset of the items

i::Int64: Index of the parents-item

ref_key::Int64: Reference Key, which can be for example the total size of the subset

ClassicAlgorithmsCollections.find_parent_in_spanning_tree — Method

find_parent_in_spanning_tree(parent::Array{Int64,1}, i::Int64, ref_key::Int64)

The find parent algorithm is a recursive function to find the subeset of an item i for graph test of having a spanning tree.

Arguments

parent::Array{Int64,1}: Array of the subset of the items

i::Int64: Index of the parents-item

ClassicAlgorithmsCollections.find_word — Method

find_word(
    graph::Dict{Int64,Array{String,1}},
    visited::Array{Bool,2},
    i::Int64,
    j::Int64,
    size_v::Int64,
    size_h::Int64,
    current_word::String,
    reference_words::Array{String,1},
    result::Array{String,1},
)

Based on the Depth First Traversal algorithm, words will be find for a current char by going through the graph up and down and keep track of the visited nodes. Important is that the travelling happens in both direction up and down and left and right. Everythign will be stacked in the visited list. If no word will be found the current_word has to be deleted.

Arguments

graph::Dict{Int64,Array{String,1}}: Graph of the connected nodes of chars, which can build the words

visited::Array{Bool,2}: List of the visited chars in the graph

i::Int64: current row

j::Int64: current col

size_v::Int64: total size of rows

size_h::Int64: total size of cols

current_word::String: current joint word of chars

reference_words::Array{String,1}: Reference words to search for

result::Array{String,1}: List of the found words in the graph

ClassicAlgorithmsCollections.graph_bridge_check — Method

graph_bridge_check(graph::Dict{Int64,Array{Int64,1}}))

For finding a bridge or more bridges in an undirect connected graph, the kind of connection has to be found, which can disconnect the graph by removing it. In case of disconnected undirected graphs, the bridge is the connection which increases number of disconnected components by removing it.

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes

Examples

julia> import ClassicAlgorithmsCollections
julia> graph_bridge = Dict(1 => [2, 3, 4], 2 => [1, 3], 3 => [1, 2], 4 => [1, 5], 5 => [4])
julia> ClassicAlgorithmsCollections.graph_bridge_check(graph_bridge)
[4 5; 1 4]

ClassicAlgorithmsCollections.graph_cycle_check — Method

graph_cycle_check(graph::Dict{Int64,Array{Int64,1}})

The disjoint-set data structure principle is used to check if a direct or undirect graph contains a cycle. For this reason, the algorithm keeps the first track of a set of items partitioned into several disjoint (non-overlapping) subsets to find which subset a particular item is kept. This procedure is essential to figure out if two items are in the same subgroup. Next, the two subsets have to be merged into a single subset. For more information see: https://en.wikipedia.org/wiki/Disjoint-setdatastructure

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

Examples

julia> import ClassicAlgorithmsCollections
julia> graph_cycle_true = Dict(1 => [2], 2 => [3], 3 => [1, 4])
julia> ClassicAlgorithmsCollections.graph_cycle_check(graph_cycle_true)
true
julia> graph_cycle_false = Dict(1 => [2], 2 => [5], 3 => [1, 4])
julia> ClassicAlgorithmsCollections.graph_cycle_check(graph_cycle_false)
false

ClassicAlgorithmsCollections.heap_sorting — Method

heap_sorting(array::Array{Int64,1})

As a comparison-based sorting algorithm, the heapsort algorithm (HSA) divides its input into a sorted and an unsorted region, and it iteratively shrinks the unsorted region by extracting the largest element from it and inserting it into the sorted region. A specialty is that the HSA keeps the unsorted region in a heap data structure to find the largest element in each step more quickly. In more detail, in the first part of the HSA (while-loop), the largest value has to be found and set to position one. In the second part of the HSA (while-loop), the array's first and largest value has to be swap to the last index of the array, and the swapping-procedure starts again for a new interval n-1. For more information see: https://en.wikipedia.org/wiki/Heapsort

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

...

Examples

julia> arr = [64, 34, 25, 12, 22, 11, 90] 
julia> ClassicAlgorithmsCollections.heap_sorting(arr)
[11, 12, 22, 25, 34, 64, 90]

ClassicAlgorithmsCollections.heapify — Method

heapify(array::Array{Int64,1})

The heapify function merges the two subroutines for: 1) Put elements of the array in heap order 2) Repair the heap elements i whose root element, if is not at the index largest For more information see: https://en.wikipedia.org/wiki/Heapsort

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

n::Int64: Length of the array

::Int64: Current index of the array

...

ClassicAlgorithmsCollections.initialize_matrices — Method

initialize_matrices(graph::Dict{Int64,Array{Tuple{Int64,Int64},1}})

Initialize the matrices for distances and pathes.

...

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes with weights

...

ClassicAlgorithmsCollections.insertion_sorting — Method

insertion_sorting(array::Array{Int64,1})

The insertion sorting algorithm builds the final sorted array by inserting elements that are greater than the key, to one position ahead of their current position step one item at a time. For more information see: https://en.wikipedia.org/wiki/Insertion_sort

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

...

Examples

julia> arr = [64, 34, 25, 12, 22, 11, 90] 
julia> ClassicAlgorithmsCollections.insertion_sorting(arr)
[11, 12, 22, 25, 34, 64, 90]

ClassicAlgorithmsCollections.interpolation_searching — Method

interpolation_searching(array::Array{Int64,1}, target::Int64)

The Interpolation search algorithm (ISA) finds a target-position in a sorted array by using a numerical procedure. The sorting procedure uses a linear fitting for finding the target position of the remaining search space in the array in more detail. The array's target position is calculated by the straight slope between the lowest and largest boundary of the remaining array and the lowest array position itself during each optimization cycle. If the target-position cannot be found, the array-space will be shrink for the lower or higher boundary region based on a comparison. For more information see: https://en.wikipedia.org/wiki/Interpolation_search

...

Arguments

array::Array{Int64,1}: Sorted array of integers

target::Int64: Target-value to find the position

...

Examples

julia> arr = [10,11, 12, 14, 16, 18, 19, 20, 21, 22, 23, 24, 33, 35, 42, 47]
julia> target = 12
julia> ClassicAlgorithmsCollections.interpolation_searching(arr, target)
3

ClassicAlgorithmsCollections.merge — Method

merge(left::Array{Int64,1}, right::Array{Int64,1}))

The merge algorithms (MA) merge the subarrays left and right to produce new sorted subarrays until there is only one subarray remaining, which will be the sorted array. For more information see: https://en.wikipedia.org/wiki/Merge_sort

...

Arguments

left::Array{Int64,1}: Unsorted left part of the array

right::Array{Int64,1}: Unsorted right part of the array

...

ClassicAlgorithmsCollections.merge_sorting — Method

merge_sorting(array::Array{Int64,1})

The merge sort algorithms (MSA) are a comparison-based sorting algorithm, which is referred to as the divide and conquer algorithms. The stable sort implementation is a widely used method for the MSA, which means that the order of equal elements is the same in the input and output. In the current implementation, a top-down implementation is used; however, a Bottom-up implementation can be used, too. In the top-down implementation, the MSA recursively splits the array into subarrays until the subarray size is < 2, merging those subarrays to produce a sorted array by using a new function merge. The back copying is blocked by alternating the direction of the merge with each recursion. For more information see: https://en.wikipedia.org/wiki/Merge_sort

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

...

Examples

julia> arr = [64, 34, 25, 12, 22, 11, 90] 
julia> ClassicAlgorithmsCollections.merge_sorting(arr)
[11, 12, 22, 25, 34, 64, 90]

ClassicAlgorithmsCollections.minimum_spanning_tree — Method

minimum_spanning_tree(graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}))

The minimum spanning tree (MST) algorithm detects a subset of the edges of a connected, edge-weighted undirected graph that connects all the vertices together. The MST algorithms focus is a) to exclude any cycles and b) to find the minimum possible total edge weight, which will create a spanning tree whose sum of edge weights is as small as possible. The Kruskal's algorithm is used to find the minimum spanning forest of an undirected edge-weighted graph. For more information see: https://en.wikipedia.org/wiki/Minimumspanningtree and https://en.wikipedia.org/wiki/Kruskal%27s_algorithm

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes with the weights

Examples

julia> import ClassicAlgorithmsCollections
julia> graph_with_spanning_tree = Dict(1 => [(2, 10), (3, 6), (4, 5)], 2 => [(4, 15)], 3 => [(4, 4)]
julia> ClassicAlgorithmsCollections.minimum_spanning_tree(graph_with_spanning_tree)
[3 4 4; 1 4 5; 1 2 10]

ClassicAlgorithmsCollections.modular_exponentiation — Method

modular_exponentiation(base::Int, exponent::Int, modulus::Int)

Compute the residuum of the base raised to the exponent, which is divided by the modulus.

...

Arguments

base::Integer: base

exponent::Integer: exponent

modulus ::Integer: modulus

...

Examples

julia> import ClassicAlgorithmsCollections
julia> ClassicAlgorithmsCollections.modular_exponentiation(2, 3, 15)
8

ClassicAlgorithmsCollections.partition — Method

partition(array::Array{Int64,1}, low::Int64, high::Int64)

The partion algorithm is shuffeling the array for a given interval of low and high boundaries.

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

low::Int64: Lowest index of the unsorted array or subarray

high::Int64: Highes index of the unsorted array or subarray

...

ClassicAlgorithmsCollections.path_reconstruction — Method

path_reconstruction(next::Array{Int64,2}, u::Int64, v::Int64)

Reconstruction of the actual path between any two endpoint vertices (u & v). ...

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes with weights

u::Int64: Startpoint of the to investigated path

v::Int64: Endpoint of the to investigated path

...

ClassicAlgorithmsCollections.quick_sorting — Function

quick_sorting(array::Array{Int64,1}, low = nothing, high = nothing)

The quick sort algorithm (QSA) works by selecting a pivot element from the array and partitioning the other elements into two subarrays, according to whether they are less than or greater than the pivot-window. Then the sorting of subarrays is recursively organized. This procedure repeatedly happens until each subarray is organized; consequently, the subarrays' merging is an organized array. For more information see: https://en.wikipedia.org/wiki/Quicksort#Parallelization

...

Arguments

array::Array{Int64,1}: Unsorted array of integers

low::Int64: Lowest index of the unsorted array or subarray

high::Int64: Highes index of the unsorted array or subarray

...

Examples

julia> arr = [64, 34, 25, 12, 22, 11, 90] 
julia> ClassicAlgorithmsCollections.quick_sorting(arr)
[11, 12, 22, 25, 34, 64, 90]

ClassicAlgorithmsCollections.shortest_path_tree — Function

shortest_path_tree(graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}, u=nothing, v=nothing)

The Shortest Path Tree (SPT) algorithm solves the shortest path problem between every pair of vertices in a given edge-weighted directed Graph based on the Floyd–Warshall algorithm. Optional, the SPT also provides the total parts between a start- (u) and end- point (v). For more information see: https://en.wikipedia.org/wiki/Floyd–Warshall_algorithm

Arguments

next::Array{Int64,2}: Vertex matrix of the connected nodes

u::Int64: Startpoint of the to investigated path; optional

v::Int64: Endpoint of the to investigated path; optional

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = Dict( 1 => [(3, -2)], 2 => [(1, 4),(3, 3)], 3 => [(4, 2)], 4 => [(2, -1)]
julia> ClassicAlgorithmsCollections.shortest_path_tree(graph, 2, 4)
([0 -1 -2 0; 4 0 2 4; 5 1 0 2; 3 -1 1 0], [2, 1, 3, 4])

ClassicAlgorithmsCollections.sortrows — Function

Provided by: https://discourse.julialang.org/t/sort-matrix-based-on-the-elements-of-a-specific-column/23475/5