Dynamic-Programming

ClassicAlgorithmsCollections.lis_length — Function

lis_length(str_1::String, str_2::String)

The longest increasing subsequence (LIS) algorithm is dedicated to finding the longest increasing subsequence for a given array. This subsequence does not have to be contiguous or unique.

1. Figure out the number of possible different subsequences of a string with length `m` and `n`.
2. Recontracted `for-loops` for defining the outer layer of the storing matrix `C`.
3. Checking if `str_1[i] == str_2[j]` and updating the count in  matrix `C`.
4.Return the latest element of `C[m,n]`.

For more information see: https://en.wikipedia.org/wiki/Longestincreasingsubsequence

Arguments

array::Array{Int64,1}: array of integers unsortetd or sorted

Examples

julia> import ClassicAlgorithmsCollections
julia> arr = [10, 22, 9, 33, 21, 50, 41, 60] 
julia> ClassicAlgorithmsCollections.lis_length(arr)
4

ClassicAlgorithmsCollections.minumum_operations — Function

minumum_operations(str_1::String, str_2::String)

The minimum operations algorithm explores how many string edits (insert, remove, replace) are needed to convert str1 to str2.

Arguments

str_1::String: String 1 of length m

str_2::String: String 2 of length n

Examples

julia> import ClassicAlgorithmsCollections
julia> str1 = "sunday"
julia> str2 = "saturday"
julia> ClassicAlgorithmsCollections.minumum_operations(str1,str2)
3

ClassicAlgorithmsCollections.number_of_steps — Function

number_of_steps(distance::Int64, step_size::Int64)

Number of stepsize is calculating the total number of combinations for walking a distance for a given number of stepsize. For example, for a distance of 3 with a maximum step_size of 3 will provide for possible combinations:

- 1 step + 1 step + 1 step
- 1 step + 2 step
- 2 step + 1 step
- 3 step

Arguments

distance::Int64: Distance to go

step_size::Int64: Maximum step size of walking a distance

Examples

julia> import ClassicAlgorithmsCollections
julia> ClassicAlgorithmsCollections.number_of_steps(8,3)
81

ClassicAlgorithmsCollections.subset_sum_test — Function

subset_sum_test(array::Array{Int64,1}, target::Int64)

The recursive expression tests if there is a subset in the array, where the sum of the subset is equal to the target.

Arguments

array::Array{Int64,1}: Array with a possible sum of a subsequence.

target::Int64: Target sum of the subsequence, which should be included in the array.

Examples

julia> import ClassicAlgorithmsCollections
julia> set = [3, 34, 4, 12, 5, 2]
julia> sum = 9
julia> ClassicAlgorithmsCollections.subset_sum_test(set, sum)
true

ClassicAlgorithmsCollections.knapsack_solver — Function

knapsack_solver(array::Array{Int64,1}, weight::Array{Int64,1}, capacity::Int64)

The knapsack problem is linked to the combinatorial optimization. In more detail all array-elements are characterized by a given weight. The excersive is to maximize the the value of subsequence of array-elements by not ooutbiding the capacity. For this purpose a recurisve comparsion between old and new added items is designed wich continously update the maximum reacahbele value sum. For more information see: https://en.wikipedia.org/wiki/Knapsack_problem

Arguments

array::Array{Int64,1}: Array with the current values of each items.

weight::Array{Int64,1}: Array with the current weoght of each items.

capacity::Int64: Maximum weight to carry.

Examples

julia> import ClassicAlgorithmsCollections
julia> val = [60, 100, 120, 230, 220]
julia> wt = [10, 20, 25, 30, 25]
julia> W = 50
julia> ClassicAlgorithmsCollections.knapsack_solver(val, wt, W)
340