Example

Example Julia package repo.

ClassicAlgorithmsCollections.breadth_first_search — Method

breadth_first_search(graph::Dict{Int64,Array{Int64,1}}, start::Int64)

The breadth-first search (BFS) is an algorithm dedicated to traversing or searching for tree or graph data structures. It starts at a specified tree root (start) for exploring all connected neighbor nodes. The important feature is that the BFS automatically leaves the present depth and passes on to the next nodes at a deeper level. BFS is queue-based. For more information see: https://en.wikipedia.org/wiki/Breadth-first_search

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

start::Int64: Startpoint (first selected vertex) of the graph-traveling process

...

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = Dict(1=> [2, 3], 2=> [3], 3=> [1, 4], 4=> [4])
julia> ClassicAlgorithmsCollections.breadth_first_search(graph, 3)
[3, 1, 4, 2]

ClassicAlgorithmsCollections.depth_first_search — Method

depth_first_search(graph::Dict{Int64,Array{Int64,1}}, start::Int64)

The depth-first search (DFS) is an algorithm dedicated to traversing or searching for tree or graph data structures. It starts at a specified tree root (start) for exploring as far as possible at each branch. Afterthat the BFS starts automatically backtracking. DFS is stack-based. For more information see: https://en.wikipedia.org/wiki/Depth-first_search

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

start::Int64: Startpoint (first selected vertex) of the graph-traveling process

...

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = Dict(1=> [2, 3], 2=> [3], 3=> [1, 4], 4=> [4])
julia> ClassicAlgorithmsCollections.breadth_first_search(graph, 3)
[3, 1, 2, 4]

ClassicAlgorithmsCollections.dfs_recrusive — Method

dfs_recrusive(
    graph::Dict{Int64,Array{Int64,1}},
    start::Int64,
    visited::Array{Bool,1},
    solution::Array{Int64,1},
)

The recursively call of the dfs_recrusive is essential for exploring each single branch of the graph.

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

start::Int64: Startpoint (first selected vertex) of the graph-traveling process

visited::Array{Bool,1}: Visited nodes of the graph

solution::Array{Int64,1}: Solution of the raph-traveling proces

...

ClassicAlgorithmsCollections.find_global_maximum — Method

find_global_maximum(graph::Dict{Int64,Array{Int64,1}})

Find the total global maximum based on the intial vertex, the dictionary key, and the accesible vertexes, the array-list.

...

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

...

ClassicAlgorithmsCollections.find_parent — Method

find_parent(parent::Array{Int64,1}, i::Int64, vertex_size::Int64)

The find parent algorithm is a recursive function to find the subeset of an item i.

Arguments

parent::Array{Int64,1}: Array of the subset of the items

i::Int64: Index of the item

vertex_size::Int64: Total size of the subeset

ClassicAlgorithmsCollections.graph_cycle_check — Method

graph_cycle_check(graph::Dict{Int64,Array{Int64,1}})

The disjoint-set data structure principle is used to check if a direct or undirect graph contains a cycle. For this reason, the algorithm keeps the first track of a set of items partitioned into several disjoint (non-overlapping) subsets to find which subset a particular item is kept. This procedure is essential to figure out if two items are in the same subgroup. Next, the two subsets have to be merged into a single subset. For more information see: https://en.wikipedia.org/wiki/Disjoint-setdatastructure

Arguments

graph::Dict{Int64,Array{Int64,1}}: Graph of the connected nodes

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = graph_cycle_true = Dict(1 => [2], 2 => [3], 3 => [1, 4])
julia> ClassicAlgorithmsCollections.graph_cycle_check(graph_cycle_true)
true
julia> graph = graph_cycle_false = Dict(1 => [2], 2 => [5], 3 => [1, 4])
julia> ClassicAlgorithmsCollections.graph_cycle_check(graph_cycle_false)
false

ClassicAlgorithmsCollections.initialize_matrices — Method

initialize_matrices(graph::Dict{Int64,Array{Tuple{Int64,Int64},1}})

Initialize the matrices for distances and pathes.

...

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes with weights

...

ClassicAlgorithmsCollections.modular_exponentiation — Method

modular_exponentiation(base::Int, exponent::Int, modulus::Int)

Compute the residuum of the base raised to the exponent, which is divided by the modulus.

...

Arguments

base::Integer: base

exponent::Integer: exponent

modulus ::Integer: modulus

...

Examples

julia> import ClassicAlgorithmsCollections
julia> ClassicAlgorithmsCollections.modular_exponentiation(2, 3, 15)
8

ClassicAlgorithmsCollections.path_reconstruction — Method

path_reconstruction(next::Array{Int64,2}, u::Int64, v::Int64)

Reconstruction of the actual path between any two endpoint vertices (u & v). ...

Arguments

graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}: Graph of the connected nodes with weights

u::Int64: Startpoint of the to investigated path

v::Int64: Endpoint of the to investigated path

...

ClassicAlgorithmsCollections.shortest_path_tree — Function

shortest_path_tree(graph::Dict{Int64,Array{Tuple{Int64,Int64},1}}, u=nothing, v=nothing)

The Shortest Path Tree (SPT) algorithm solves the shortest path problem between every pair of vertices in a given edge-weighted directed Graph based on the Floyd–Warshall algorithm. Optional, the SPT also provides the total parts between a start- (u) and end- point (v). For more information see: https://en.wikipedia.org/wiki/Floyd–Warshall_algorithm

Arguments

next::Array{Int64,2}: Vertex matrix of the connected nodes

u::Int64: Startpoint of the to investigated path; optional

v::Int64: Endpoint of the to investigated path; optional

Examples

julia> import ClassicAlgorithmsCollections
julia> graph = Dict( 1 => [(3, -2)], 2 => [(1, 4),(3, 3)], 3 => [(4, 2)], 4 => [(2, -1)]
julia> ClassicAlgorithmsCollections.shortest_path_tree(graph, 2, 4)
([0 -1 -2 0; 4 0 2 4; 5 1 0 2; 3 -1 1 0], [2, 1, 3, 4])